Refactoring the formatting to options & changing printSuccess to resultsFormat
The formatting options are really going to be a list of options for the object,
so I changed that around.

My tests were checking the result expecting the documentation format. The
printSuccesses option was actually a flag for that format, so I refactored that
code and the tests of those as well.

Refactored the tests to not use lspec_clearSuite (unless testing it), but to
instead use lspec_setupForTesting which is defined at the top of the tests code
and is used to completely reset up lspec including the options I'll want as the
defaults for testing.

TODO:
I need to add more logic to the resultsFormat method to check for valid formats
being passed to it. (Of course, I need to add the tests first.)

I also need to add the tests and logic for the 'p' output format which I want
to be the default. (p stands for progress - d is what I use for tests and it
stands for documentation.)