Added the beginnings of [expect->error]
Basic tests for this method are passing:
  - Block raises the correct error
  - Block doesn't raise an error
  - Doesn't get passed a block

ToDo:
What happens when a block gets an unexpected error?

We also need to put in some logic for what happens when there's an unexpected
failure. We should probably gracefully catch it with its call stack and mark
the test as failing. (More tests needed for this.)