Adding [expect(expect, got)] method for better diagnostics messages
When [expect(exepect == got)] fails, you get "expected: true" and "got: false"
which isn't the greatest message. The new [expect(expect, got)] compares for
equality and if it doesn't find it, then the first value is the "expected" one
and the second is the "got" value in the failure message.

I have also updated the lspec command test spec to use this method.