Changing the way locals are handled & updating tests
Example Code
============
describe('something') => {
    with item in #my_map do {
      local(my_val) = #item->find('value')
        describe('in some context') => {
          it('does something') => {
            #my_val
           }
       }
   }
}
============
[describe] and [it] take their [givenBlock] and add it to one of an
[lspec_testSuite] object's internal data members. The object later invokes
those blocks in a prescribed order. Since that is after the query expression
has finished iterating, [#my_val] is set to the last item in the iteration
(and so is [#item]).

I changed the code to store a deep copy of the [givenBlock] instead, and this
fixes the problem - the locals keep the values from when the method was called.

I haven't decided whether or not [beforeXXX] and [afterXXX] need this too.